sender/requirements.txt
flask
paho-mqtt
numpy